# src/main/java/dev/onepieceapi/userservice/adapter/in/web/AuditController.java
package dev.onepieceapi.userservice.adapter.in.web;

import dev.onepieceapi.userservice.adapter.in.web.dto.AuditEventResponse;
import dev.onepieceapi.userservice.adapter.in.web.dto.PageResponse;
import dev.onepieceapi.userservice.adapter.in.web.mapper.AuditEventResponseMapper;
import dev.onepieceapi.userservice.application.service.AuditQueryService;
import dev.onepieceapi.userservice.domain.AuditAction;
import dev.onepieceapi.userservice.domain.AuditEvent;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springdoc.core.annotations.ParameterObject;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.time.LocalDate;
import java.util.List;
import java.util.Set;
import java.util.UUID;

/**
 * The audit read path from Step 17 - {@code GET /audit}, gated by the {@code audit:read}
 * permission authority (see {@code security.SecuredEndpoint}). Kept out of
 * {@link UserController}: it reads from
 * {@link dev.onepieceapi.userservice.application.port.out.AuditLogPort}, not
 * {@code UserDirectoryPort}.
 */
@RestController
@Tag(name = "Audit")
@RequiredArgsConstructor(onConstructor_ = { @Autowired })
class AuditController {

	private final AuditQueryService auditQueryService;

	@GetMapping(ApiPaths.AUDIT)
	PageResponse<AuditEventResponse> listAuditEvents(@ParameterObject Pageable pageable,
			@RequestParam(required = false) UUID userId, @RequestParam(required = false) Set<AuditAction> actions,
			@RequestParam(required = false) String actorEmail, @RequestParam(required = false) LocalDate from,
			@RequestParam(required = false) LocalDate to) {
		Page<AuditEvent> page = this.auditQueryService.list(pageable, userId, actions, actorEmail, from, to);
		return PageResponse.from(page.map(AuditEventResponseMapper::toResponse));
	}

	/**
	 * Every actor who has ever recorded an event, sorted - powers the Ship's Log author
	 * filter.
	 */
	@GetMapping(ApiPaths.AUDIT_ACTORS)
	List<String> listActors() {
		return this.auditQueryService.listActors();
	}

}